backend_app/pyproject.toml
[project]
name = "backend-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "fastapi[all]>=0.121.0",
    "uvicorn[standard]>=0.38.0",
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.0.1",
    "langgraph>=1.0.2",
    "langchain>=1.0.3",
    "asyncpg>=0.30.0",
    "passlib[bcrypt]==1.7.4",
    "pyjwt==2.9.0",
    "pydantic[email]>=2.12.3",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
]
